app/requirements.txt
python-telegram-bot>=20.0
tinkoff-investments
pydantic-settings>=2.0
redis>=4.5.4
asyncpg>=0.28.0
apscheduler>=3.10.0
aiohttp>=3.8.0
async-timeout>=4.0.2
backoff==2.2.1
pydantic>=2.0
numpy>=1.22.0
# Новые зависимости для webhook сервера
gunicorn>=21.2.0
uvloop>=0.17.0  # Для лучшей производительности
certifi>=2023.5.7  # Для SSL сертификатов